Invoke ActiveSupport::Dependencies.clear after test runs
On Rails 4.1+, depending on the order the tests run in, the Book class
created during test setup will sometimes be resolved to the same class
used by a previous test method, even though the constant it was assigned
to has been unset. This is because the class is still assigned
anonymously as the value to the "Book" key in an instance of
ActiveSupport::Dependencies::ClassCache.

You can see this symptom by checking out the previous commit and
running:

  BUNDLE_GEMFILE=gemfiles/Rails_42.gemfile bundle exec ruby  -e 'require "./test.rb"' -- --seed 43 -n "/^(?:DefaultValuePluginTest#(?:test_works_with_nested_attributes|test_can_specify_default_value_via_association))$/"

Explicitly clearing the cache resolves this issue.

See #47
See #46